Created a portfolio application
I would mark this as "finished", but since I have yet to make any of the applications,
it's not really finished. In addition, I haven't done any styling. So, I guess this is
just an initial commit.

Oh, and here is a list of sites that I used to help me:
* <http://stackoverflow.com/questions/7326155/align-text-center-with-android>
* <https://developer.android.com/reference/android/support/design/widget/Snackbar.html>
* <https://developer.android.com/guide/topics/resources/string-resource.html>
* <http://stackoverflow.com/questions/4992357/how-to-load-my-app-from-eclipse-to-my-android-phone-instead-of-avd>
* <http://stackoverflow.com/questions/15152933/android-how-to-add-a-custom-xml-file-in-res-values-and-how-to-register-the-custo>
* <http://stackoverflow.com/questions/7666589/using-getresources-in-non-activity-class>
* <http://www.vogella.com/tutorials/AndroidListView/article.html>

Not everything that I looked up made it into the app.